Moved julia import test
