+++ test/JuliaImport/Fujita/measurementData_Fujita_SciSignal2010.tsv
